Move components contenxt to core
