+++ packages/toolpad-core/src/componentsContext.tsx
@@ -1,0 +1,7 @@
+import { ToolpadComponents } from './types';
+import { createProvidedContext } from './utils/react';
+
+const [useComponents, ComponentsContextProvider] =
+  createProvidedContext<ToolpadComponents>('Components');
+
+export { useComponents, ComponentsContextProvider };
